+++ app/build.gradle.kts
@@ -48,6 +48,7 @@ dependencies {
     implementation(project(":core:designsystem"))
     implementation(project(":core:model"))
     implementation(project(":core:common"))
+    implementation(project(":core:ui"))
 
     implementation(libs.core.ktx)
     implementation(libs.activity.compose)
