+++ app/src/main/java/cord/eoeo/momentwo/ui/photodetail/composable/CommentBottomSheet.kt
@@ -66,8 +66,8 @@ import androidx.compose.ui.unit.sp
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.itemKey
 import coil.ImageLoader
-import cord.eoeo.momentwo.ui.composable.CircleAsyncImage
-import cord.eoeo.momentwo.ui.composable.TextFieldDialog
+import cord.eoeo.momentwo.core.ui.CircleAsyncImage
+import cord.eoeo.momentwo.core.ui.TextFieldDialog
 import cord.eoeo.momentwo.core.model.CommentItem
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalLayoutApi::class)
