+++ core/ui/src/main/java/cord/eoeo/momentwo/core/ui/UserItemBox.kt
@@ -1,4 +1,4 @@
-package cord.eoeo.momentwo.ui.composable
+package cord.eoeo.momentwo.core.ui
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
